Force 2D rendering in TransformJS
This avoids an obscure sporadic rendering bug in Chrome where cards
turned over onto the waste are not rendered face-up, even though their
background-position has been updated. Hovering the elements in the
Developer Tools (thereby forcing a repaint) makes the card faces appear,
so it appears to be a genuine issue with the rendering engine. Forcing
2D so far seems to make this go away.